Add logo for ZoZo Token
ZoZo is a meme token designed to support the community of meme coins. With unique tokenomics and an engaging story, ZoZo aims to become the last savior of meme coins in the crypto market. The project combines fun, community-driven goals with a strong focus on sustainable growth and value creation for holders."